requirements.txt
azure-ai-projects==1.0.0b11
azure-ai-agents==1.1.0b2
azure-identity==1.13.0
python-dotenv==1.0.1
